--- src/App/activity.resolver.ts
@@ -1,157 +1,0 @@
-import { UseInterceptors } from '@nestjs/common'
-import {
-  Args,
-  ArgsType,
-  Field,
-  Int,
-  Parent,
-  Query,
-  ResolveField,
-  Resolver,
-} from '@nestjs/graphql'
-import { Connection } from 'typeorm'
-import { ActivityType } from './utils/queryHelpers';
-import Activity from '../Database/Entities/activity.entity'
-import { Author } from '../Database/Entities/types/IUser'
-import { IWiki } from '../Database/Entities/types/IWiki'
-import SentryInterceptor from '../sentry/security.interceptor'
-import PaginationArgs from './pagination.args'
-
-@ArgsType()
-class ActivityArgs extends PaginationArgs {
-  @Field(() => String, { nullable: true })
-  wikiId!: string
-
-  @Field(() => String)
-  lang = 'en'
-
-}
-
-@ArgsType()
-class ActivityArgsByUser extends PaginationArgs {
-  @Field(() => String)
-  userId!: string
-}
-@ArgsType()
-class ActivityByCategoryArgs extends PaginationArgs {
-  @Field(() => ActivityType)
-  type = ActivityType.CREATED
-
-  @Field(() => String)
-  category!: string
-}
-
-@ArgsType()
-class ByIdAndBlockArgs extends ActivityArgs {
-  @Field(() => Int)
-  block!: number
-}
-
-@UseInterceptors(SentryInterceptor)
-@Resolver(() => Activity)
-class ActivityResolver {
-  constructor(private connection: Connection) {}
-
-  @Query(() => [Activity])
-  async activities(@Args() args: ActivityArgs) {
-    const repository = this.connection.getRepository(Activity)
-    return repository
-      .createQueryBuilder('activity')
-      .leftJoin('wiki', 'w', 'w."id" = activity.wikiId')
-      .leftJoinAndSelect('activity.user', 'user')
-      .where(`activity.language = '${args.lang}' AND w."hidden" = false`)
-      .cache(
-        `activities_cache_limit${args.limit}-offset${args.offset}-lang${args.lang}`,
-        60000,
-      )
-      .limit(args.limit)
-      .offset(args.offset)
-      .orderBy('datetime', 'DESC')
-      .getMany()
-  }
-
-  @Query(() => [Activity])
-  async activitiesByWikId(@Args() args: ActivityArgs) {
-    const repository = this.connection.getRepository(Activity)
-    return repository
-      .createQueryBuilder('activity')
-      .leftJoin('wiki', 'w', 'w."id" = activity.wikiId')
-      .where(`activity.wikiId = '${args.wikiId}' AND w."hidden" = false`)
-      .limit(args.limit)
-      .offset(args.offset)
-      .orderBy('datetime', 'DESC')
-      .getMany()
-  }
-
-  @Query(() => [Activity])
-  async activitiesByCategory(@Args() args: ActivityByCategoryArgs) {
-    const repository = this.connection.getRepository(Activity)
-    return repository
-      .createQueryBuilder('activity')
-      .leftJoin('wiki', 'w', 'w."id" = activity.wikiId')
-      .leftJoin(
-        'wiki_categories_category',
-        'c',
-        'c."categoryId" = :categoryId',
-        {
-          categoryId: args.category,
-        },
-      )
-      .where(`c."wikiId" = activity.wikiId AND  w."hidden" = false AND type = '${args.type}'`)
-      .limit(args.limit)
-      .offset(args.offset)
-      .orderBy('datetime', 'DESC')
-      .getMany()
-  }
-
-  @Query(() => [Activity])
-  async activitiesByUser(@Args() args: ActivityArgsByUser) {
-    const repository = this.connection.getRepository(Activity)
-    return repository
-      .createQueryBuilder('activity')
-      .leftJoin('wiki', 'w', 'w."id" = activity.wikiId')
-      .where(` activity.userId = :user AND w."hidden" = false`, {
-        user: args.userId,
-      })
-      .orderBy('activity.datetime', 'DESC')
-      .limit(args.limit)
-      .offset(args.offset)
-      .getMany()
-  }
-
-  @Query(() => Activity)
-  async activityById(@Args('id', { type: () => String }) id: string) {
-    const repository = this.connection.getRepository(Activity)
-    return repository
-      .createQueryBuilder('activity')
-      .leftJoin('wiki', 'w', 'w."id" = activity.wikiId')
-      .where(`activity.id = '${id}' AND w."hidden" = false`)
-      .getOne()
-  }
-
-  @Query(() => Activity)
-  async activityByWikiIdAndBlock(@Args() args: ByIdAndBlockArgs) {
-    const repository = this.connection.getRepository(Activity)
-    return repository
-      .createQueryBuilder('activity')
-      .leftJoin('wiki', 'w', 'w."id" = activity.wikiId')
-      .where(`activity.wikiId = '${args.wikiId}' AND w."hidden" = false`)
-      .andWhere(
-        `activity.language = '${args.lang}' AND activity.block = '${args.block}'`,
-      )
-      .getOne()
-  }
-
-  @ResolveField(() => Author)
-  async author(@Parent() wiki: IWiki) {
-    const { id } = wiki
-    const repository = this.connection.getRepository(Activity)
-    const res = await repository.query(`SELECT "userId", u.* 
-        FROM activity
-        LEFT JOIN "user_profile" u ON u."id" = "userId"
-        WHERE "id" = '${id}' AND "type" = '0'`)
-    return { id: res[0]?.userId, profile: { ...res[0] } || null }
-  }
-}
-
-export default ActivityResolver
