+++ src/App/pinJSONAndImage/pin.resolver.spec.ts
@@ -10,11 +10,11 @@ import * as fs from 'fs/promises'
 import { HttpModule } from '@nestjs/axios'
 import PinService from './pin.service'
 import PinResolver from './pin.resolver'
-import ActivityService from '../activity.service'
 import IPFSValidatorService from '../../Indexer/Validator/validator.service'
 import PinJSONErrorWebhook from './webhookHandler/pinJSONErrorWebhook'
 import MetadataChangesService from '../../Indexer/Store/metadataChanges.service'
 import WebhookHandler from '../utils/discordWebhookHandler'
+import ActivityService from '../Activities/activity.service'
 
 jest.mock('fs')
