+++ src/Relayer/relayer.module.ts
@@ -4,7 +4,7 @@ import RelayerService from './services/relayer.service'
 import RelayerController from './controllers/relayer.controller'
 import RelayerResolver from './resolvers/relayer.resolver'
 import httpModule from '../httpModule'
-import ActivityService from '../App/activity.service'
+import ActivityService from '../App/Activities/activity.service'
 
 @Module({
   imports: [httpModule(10000)],
